:recycle: Types 폴더 생성 및 타입 분리
